Remove the possibility of "@id" to become the @type value
The new model doesn’t allow the template field to denote an object
identifier, such that having a pair “@type” : “@id” is invalid. This
construction is only valid within the @context object to denote the
value type of the declared property is an object identifier (i.e., IRI)

The value for the type field can only be either a URI string or an
array of URI strings. The value will reference to a data type (e.g.,
XSD types)